requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.0.1
Flask-FlatPages==0.6
Frozen-Flask==0.11
Jinja2==2.7.3
Markdown==2.6.1
MarkupSafe==0.23
Pillow==2.7.0
PyYAML==3.11
Werkzeug==0.10.1
itsdangerous==0.24
-e git://github.com/ryancox/motionless.git@24b8aa27b56c6e2fe8c9c6e6103a3742d6da0e0e#egg=motionless-master
requests==2.5.3
wsgiref==0.1.2
